[WIP] Apollo Guidance Computer disassembly (#54)
* Implemented basic structure for AGC plugin.

Implemented integration into build system and layout, as well as some
barebones setup code. Now onto the actual logic :3

* Added basic structure of AGC disassembler

The core disassembly logic has been ported over from
https://github.com/alfonsokim/apollo-guidance-computer/blob/master/yaAGC/agc_disassembler.c .
However, since the disassembler is inherently stateful, we still need
to implement either state or port over the whole lib / rewrite it.
Or I get to find a different solution for this.

* Done some research and gathered some information

It seems clear how to get the current address etc. from the data
passed to the plugin without resorting to internal state. That way
a lot of complexity should be avoidable and implementation gets a
lot easier.

* Cleaned up logic

* Integrated things

* Finished integration of disasm plugin.

* Worked through the AGC instruction set

Gathered info needed for the anal plugin and commented the disassembler
components.

* Fixed compiler error

* Added some constants and fixed errors

* Added help file