MDMP v1.0 (#79)
* Rework of the mdmp plugin

Major rework of the mdmp plugin, sections will load using their virtual
adressess and physical addresses will map to the correct locations in
the mdmp file if they exist (full dumps). The memory section is used to
store all memeory segments found within the mdmp file, again physical
addressess will point to the correct locations withing the file.

* Applied r2 code style guidelines

Applied the r2 code style guidelines to the code:
https://github.com/radare/radare2/blob/master/CONTRIBUTING.md

* More guideline conforming

A little more work on conforming to guidelines.

* Even more guideline conforming

Further minor changes to the coding style.

* Fixed tab spaces on bin_mdmp.c

How did I miss this file during the conversion!!!

* Fixed linking between v and paddr; added mem_info

Fixed the issue where vaddr would not link to the paddr contents. Added
initial memory info parsing support; this feature is currently disabled
as it breaks the vaddr paddr mapping.

* Added TODOs

Added TODOs to the parsing section to highlight which streams are in
need of further implementation.

* Bug fix in mdmp.c

Fixed a bug introduced with the srwx function.

* Added memory permissions to sections

I was being a bit of an idiot and have reenabled memory permissions as
this was not the cause of the vaddr/paddr issue.

* Entrypoint support

Entrypoint support for PE files has been added. Due to the fact that the
implementation is in extras the pe folder had to be cloned. This will be
removed when the plugin migrates into main.

* Fix for the entrypoint maths

Fixed an incorrect assumption about calculating entrypoints.

* Preliminary symbol support

Parses most cases of symbols, some cases are yet to be caught. Cleaning
is required before PR. Furthermore a bug fix within pe.c is required for
symbol parsing to complete successfully. Would not be surprised if there
are additional issues in pe.c!

* Initial import support

Will load imports into the import section, need to test against some
other files to double check paddr/vaddr.

* Added initial relocs support

This has added support for the relocs section. This will change due to
the leak I have introduced due to the cross dependency. As a result I
will probably now parse all modules once and store them in an RList.

* Architecture change

The first architecture change to deal with the fact that r2 uses two
binary plugins for pe files. This fixes some of my oversights and solves
a large number of issues I had reported previously.

* New architecture complete

The new more modular architecture is complete. I took some/a lot of
inspiration from nibble.

* Cleanup and some other minor changes

Including the above, had to disable the symbol function as sending bins
to pe.c is causing a seg fault.

* Fixed some memory issues

Fixed some issues with rlists assumptions. Made some quick hacky fixes
so that imports work again for now.

* Added linked library support

Will pull the linked libraries from the embedded binaries.

* Basic string support

Will now allow searching of sections for strings, i.e. enabled iz.

* String refinement

Increased the minstrlen otherwise we get string in text sections
everywhere. This could be greatly improved but identifying individual
sections!

* Re-enabled mdmp in the travis yml

Re-enabled mdmp in the travis yml and added the regression args.