Minor Makefile changes (#82)
* Renaming in Makefile.acr

The make command for building binary plugins used to be dlang which does
not make any sense, have renamed this to bin.

* Changed the mdmp.mk to prevent error

Copied the format in pyc.mk and this seems to build using the commands
in the r2pm script. So hopefully travis and r2pm will be happy now.